Move SwitchToExternalMod to its own class
